+++ src/widgets/Button.cpp
@@ -74,7 +74,7 @@ namespace pTK
             if ((wSize.width > bSize.width) || (wSize.height > bSize.height))
             {
                 // Apparently, this looks good.
-                int dMargin = (m_borderSize*2);
+                int dMargin = (int)(m_borderSize*2.0f);
                 int hMargin = ((bSize.height - wSize.height) > dMargin) ? 0 : dMargin;
                 int wMargin = ((bSize.width - wSize.width) > dMargin) ? 0 : dMargin;
                 wSize.width = ((wSize.width > bSize.width) ? wSize.width : bSize.width) + (wMargin);
